Fix database migrations sql syntax

+++ migrations/sqls/20220109181806-order-products-up.sql
@@ -1,8 +1,8 @@
 /* Replace with your SQL commands */
 CREATE TABLE order_products(
-    order_id INT,
-    product_id INT,
-    quantity INT,
+    order_id BIGINT,
+    product_id BIGINT,
+    quantity INTEGER,
     FOREIGN KEY(order_id) REFERENCES orders(id),
     FOREIGN KEY(product_id) REFERENCES products(id),
     PRIMARY KEY(order_id, product_id)
